🐛 fix(patch): dont use fork-ts-checker-webpack-plugin
