--- sources/@roots/bud-typescript/src/env.ts
@@ -1,27 +1,0 @@
-import type * as Build from '@roots/bud-framework/services/build'
-
-import type BudTypeScript from './extension.js'
-import type BudTypeCheckPlugin from './typecheck/index.js'
-
-declare module '@roots/bud-framework' {
-  interface Bud {
-    typescript: BudTypeScript
-  }
-
-  interface Modules {
-    '@roots/bud-typescript': BudTypeScript
-    '@roots/bud-typescript/typecheck': BudTypeCheckPlugin
-  }
-
-  interface Loaders {
-    ts: Build.Loader
-  }
-
-  interface Items {
-    ts: Build.Item
-  }
-
-  interface Rules {
-    ts: Build.Rule
-  }
-}
