+++ sources/@roots/bud-typescript/src/index.ts
@@ -13,7 +13,32 @@
  * @packageDocumentation
  */
 
-import './env.js'
+import type {Item, Loader, Rule} from '@roots/bud-framework/services/build'
 
 import BudTypeScript from './extension.js'
+import type BudTypeCheckPlugin from './typecheck/index.js'
+
+declare module '@roots/bud-framework' {
+  interface Bud {
+    typescript: BudTypeScript
+  }
+
+  interface Modules {
+    '@roots/bud-typescript': BudTypeScript
+    '@roots/bud-typescript/typecheck': BudTypeCheckPlugin
+  }
+
+  interface Loaders {
+    ts: Loader
+  }
+
+  interface Items {
+    ts: Item
+  }
+
+  interface Rules {
+    ts: Rule
+  }
+}
+
 export default BudTypeScript
