+++ sources/@roots/bud-typescript/src/typecheck/index.ts
@@ -18,6 +18,7 @@ import type {ForkTsCheckerWebpackPluginOptions as Options} from 'fork-ts-checker
       semantic: true,
       syntactic: true,
     },
+    mode: 'readonly',
   },
 })
 @when(async () => false)
@@ -25,12 +26,32 @@ export default class BudTypeCheckPlugin extends Extension<
   Options,
   Plugin
 > {
-  @bind
-  public async init() {
+  /**
+   * Enable typechecking
+   *
+   * @override {@link Extension.enable}
+   */
+  @bind public async enable(state: boolean = true) {
+    this.app.extensions
+      .get('@roots/bud-typescript')
+      .setOption('transpileOnly', !state)
+
+    this.logger.log(
+      `transpileOnly set to ${this.app.extensions
+        .get('@roots/bud-typescript')
+        .getOption('transpileOnly')}`,
+    )
+  }
+
+  @bind public async init() {
     const typescriptPath = await this.resolve('typescript')
 
     this.setOptions(options => ({
       ...options,
+      logger: {
+        log: this.logger.log,
+        error: this.logger.error,
+      },
       typescript: {
         ...(options.typescript ?? {}),
         typescriptPath,
